+++ backend/app/Http/Controllers/GoalController.php
@@ -29,7 +29,7 @@ class GoalController extends Controller
 		if ($number !== 3){
 			return view('goals.create');
 		} else {
-			return redirect()->route('mypage.index')->with([
+			return redirect()->route('mypage.show', ['id' => Auth::user()->id])->with([
 				'flash_message' => '同時に登録できる目標は3つまでです。',
 				'color' => 'danger'
 			]);
@@ -41,8 +41,7 @@ class GoalController extends Controller
 		$goal->fill($request->all());
 		$goal->user_id = $request->user()->id;
 		$goal->save();
-		return redirect()->route('mypage.show', [
-			'id' => Auth::user()->id,
+		return redirect()->route('mypage.show', ['id' => Auth::user()->id])->with([
 			'flash_message' => '目標を登録しました。',
 			'color' => 'success'
 		]);
@@ -50,14 +49,21 @@ class GoalController extends Controller
 
 	public function edit(Goal $goal)
 	{
-		return view('goals.edit', ['goal' => $goal]);
+		if ($goal->status === 0){
+			return view('goals.edit', ['goal' => $goal]);			
+		} else {
+			return redirect()->route('mypage.show', ['id' => Auth::user()->id])->with([
+				'flash_message' => 'クリア済みの目標は編集できません',
+				'color' => 'danger'
+			]);			
+		}
+
 	}
 
 	public function update(GoalRequest $request, Goal $goal)
 	{
 		$goal->fill($request->all())->save();
-		return redirect()->route('mypage.show', [
-			'id' => Auth::user()->id,
+		return redirect()->route('mypage.show', ['id' => Auth::user()->id])->with([
 			'flash_message' => '目標を編集しました。',
 			'color' => 'success'			
 		]);
@@ -65,12 +71,18 @@ class GoalController extends Controller
 
 	public function destroy(Goal $goal)
 	{
-		$goal->delete();
-		return redirect()->route('mypage.show', [
-			'id' => Auth::user()->id,
-			'flash_message' => '目標を削除しました。',
-			'color' => 'success'			
-		]);
+		if ($goal->status === 0){
+			$goal->delete();
+			return redirect()->route('mypage.show', ['id' => Auth::user()->id])->with([
+				'flash_message' => '目標を削除しました。',
+				'color' => 'success'			
+			]);
+		} else {
+			return redirect()->route('mypage.show', ['id' => Auth::user()->id])->with([
+				'flash_message' => 'クリア済みの目標は削除できません',
+				'color' => 'danger'
+			]);			
+		}
 	}		
 
 	public function show(Goal $goal)
